# Use ByteBuffer wrapper instead of LSB protocol