This PR is the first step of #289.

It replaces the use of `ByteBuffer` as an LSB protocol implementation with a new thin wrapper utility for `ByteBuffer`.

The modification policy here is to use the bare `ByteBuffer` methods whenever possible. The new wrapper namespace only collects functions that do what a single method of `ByteBuffer` cannot do. Most of the implementation code comes from [the implementation of the LSB protocol](https://github.com/chrovis/cljam/blob/14e0deb16f4423686d5d68907e040fe9729d262d/src/cljam/io/util/lsb.clj#L36-L82).