@athos 2023-11-15
Added accidentally omitted docstrings.