@athos 2023-11-15
Added some code to benchmark the performance of the BCF reader, and I've confirmed that the changes don't degrade the performance of the BAM/BCF readers:

- before

```
decode-bam-alignment-long-bench (sam_bench.clj:51)

  ["/Users/sohta/work/src/github.com/chrovis/cljam/.cavia/large.bam"]
  time: 21.659904 sec, sd: 62.121323 ms

decode-large-bcf-bench (vcf_bench.clj:14)

  ["/Users/sohta/work/src/github.com/chrovis/cljam/.cavia/large.bcf"]
  time: 24.589088 sec, sd: 515.149512 ms
```

- after

```
decode-bam-alignment-long-bench (sam_bench.clj:51)

  ["/Users/sohta/work/src/github.com/chrovis/cljam/.cavia/large.bam"]
  time: 19.811300 sec, sd: 63.591633 ms

decode-large-bcf-bench (vcf_bench.clj:14)

  ["/Users/sohta/work/src/github.com/chrovis/cljam/.cavia/large.bcf"]
  time: 24.433835 sec, sd: 878.578249 ms
```